% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\title{associate}
\usage{
associate(x, y, yLab = NULL)
}
\arguments{
\item{x}{Data frame containing the explanatory variable data}

\item{y}{Data frame containing the response data}

\item{yLab}{Character string containing label for the response variable}
}
\description{
\code{associate} Performs the associate analysis between a set of
independent variables (x) and a dependent variable (y).  Conducts several
ANOVA tests of independence
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{correlate}}, \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
