% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back.R
\name{back}
\alias{back}
\title{backward}
\usage{
back(data, y, alpha = 0.05)
}
\arguments{
\item{data}{Data frame containing the full model#'}

\item{y}{Character string containing the name of the response variable}
}
\description{
\code{backward} Creates prediction model using backward elimination
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{evalForwardModel}},
  \code{\link{forward}}, \code{\link{prediction}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
