% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back.R
\name{backwardElimination}
\alias{backwardElimination}
\title{backwardElimination}
\usage{
backwardElimination(data, y, predictors)
}
\description{
\code{backwardElimination} Performs step-wise backward elimination of univariate
linear regression models.  From the full model, each variable is removed
from the model, one-at-a-time, and the model is re-evaluated. The model
with the highest adjusted r-squared is retained.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{createDependent}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
