% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/challenger.R
\name{challenger}
\alias{challenger}
\title{challenger}
\usage{
challenger()
}
\value{
Data frame accuracy results for each variable added
}
\description{
\code{challenger} Iterates through correlated variables, runs all models,
and compiles results
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{createDependent}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
