% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusion.R
\name{conclusion}
\alias{conclusion}
\title{conclusion}
\usage{
conclusion(slr, mlr, data)
}
\arguments{
\item{slr}{The model object for the simple linear regresssion}

\item{mlr}{The model object for the multiregression}

\item{deta}{<- Data frame containing training data}
}
\value{
List containing data frames of summary statistics
}
\description{
\code{conclusion} Prepares summary data for conclusion section
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
