% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{correlate}
\usage{
correlate(x, y, yLab = NULL, plot = FALSE)
}
\arguments{
\item{x}{Data frame containing the explanatory variable data}

\item{y}{Data frame containing the response data}

\item{yLab}{Character string containing label for the response variable}

\item{plot}{Logical indicating whether to include a correlation plot}
}
\description{
\code{correlate} Performs the correlate analysis between a set of
independent variables (x) and a dependent variable (y).  Conducts a
correlation test and renders correlation and scatter plots.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
