% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDependent.R
\name{createDependent}
\alias{createDependent}
\title{createDependent}
\usage{
createDependent(data, mod)
}
\arguments{
\item{data}{Data frame containing movie data}

\item{mod}{= Linear model used to create the dependent variable}
}
\value{
Data frame with dependent variable added.
}
\description{
\code{createDependent} Creates a variable based upon the coefficients of
the model parameter.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{challenger}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{evalForwardModel}},
  \code{\link{forward}}, \code{\link{prediction}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
