% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{evalForwardModel}
\alias{evalForwardModel}
\title{evalForwardModel}
\usage{
evalForwardModel(data, model, y, remaining)
}
\arguments{
\item{data}{Data frame containing the full model}

\item{model, }{Data frame containing the variables and adjusted R2 values for the best model as of..}

\item{y}{Character string containing the name of the response variable}

\item{remaining}{Vector containing column names for the remaining predictors}
}
\value{
Single row data frame containing the variable and adjusted R2 for the best model evaluated
}
\description{
\code{evalForwardModel} Evaluates linear models for all remaining predictors, one
at a time, then returns the variable and adjusted R2 score for the best
model.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{forward}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
