% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullModel.R
\name{fullModel}
\alias{fullModel}
\title{fullModel}
\usage{
fullModel(data, associations, correlations)
}
\arguments{
\item{data}{Data set containing the unfiltered full model}

\item{associations}{Data frame containing associations between the categorical variables and the dependent variable.}

\item{correlations}{Data frame containing correlations between the quantitative variables and the dependent variable.}
}
\value{
Data frame listing variables selected for the full model.
}
\description{
\code{fullModel} Summarizes the variables used in the full model
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{correlate}},
  \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{evaluate}},
  \code{\link{forward}}, \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
