% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBar}
\alias{plotBar}
\title{plotBar}
\usage{
plotBar(data, yLab, xLab, plotTitle = NULL)
}
\arguments{
\item{data}{Data frame or vector containing a single categorical factor variable}

\item{yLab}{Capital case character string describing the y variable}

\item{xLab}{Capital case character string containing the name of the variable x variable}

\item{plotTitle}{Capital case character string for the title of the plot}
}
\value{
Bar plot
}
\description{
\code{plotBar} Renders a bar plot
}
\seealso{
Other visualization functions: \code{\link{plotBox}},
  \code{\link{plotCooks}}, \code{\link{plotCorr}},
  \code{\link{plotFreqProp}}, \code{\link{plotHist}},
  \code{\link{plotLinear}}, \code{\link{plotLine}},
  \code{\link{plotQQ}}, \code{\link{plotResAll}},
  \code{\link{plotResFit}}, \code{\link{plotResLeverage}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
