% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCorr}
\alias{plotCorr}
\title{plotCorr}
\usage{
plotCorr(mod, yVar)
}
\arguments{
\item{mod}{Linear model object}

\item{yVar}{Character string for the dependent variable}
}
\value{
list containing ressidual v predictor plots
}
\description{
\code{plotCorr} Creates correlation plot
}
\seealso{
Other visualization functions: \code{\link{plotBox}},
  \code{\link{plotCooks}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotLinear}},
  \code{\link{plotQQ}}, \code{\link{plotResAll}},
  \code{\link{plotResFit}}, \code{\link{plotResLeverage}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
