% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotQQ}
\alias{plotQQ}
\title{plotQQ}
\usage{
plotQQ(data, xLab = NULL, yLab, plotTitle = NULL)
}
\arguments{
\item{data}{Data frame or vector containing a single numeric variable}

\item{xLab}{Capital case character string containing the name of the grouping or subset variable (optional)}

\item{yLab}{Capital case character string containing the name of the y variable}

\item{plotTitle}{Capital case character string for the title of the plot}
}
\value{
List containing a summary statistics and QQ plot
}
\description{
\code{plotQQ} Renders a histogram with a normal curve as well as a table
of summary statistics.
}
\seealso{
Other visualization functions: \code{\link{plotBar}},
  \code{\link{plotBox}}, \code{\link{plotCooks}},
  \code{\link{plotCorr}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotLinear}},
  \code{\link{plotLine}}, \code{\link{plotResAll}},
  \code{\link{plotResFit}}, \code{\link{plotResLeverage}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
