% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{prediction}
\usage{
prediction(mods, test)
}
\arguments{
\item{test}{Data frame containing the movies data set}

\item{models}{List containing linear models used to perform predictions}
}
\value{

}
\description{
\code{prediction} Performs prediction on test data and returns results
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{correlate}},
  \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
