% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(movies, fin1, fin2)
}
\arguments{
\item{movies}{Data frame containing the movies data set}

\item{fin1}{Data set containing movie financial information}

\item{fin2}{Data set containing movie financial information}
}
\description{
\code{preprocess} Performs preprocessing of data for analysis
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{challenger}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{createDependent}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{prediction}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
