% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{process}
\usage{
process(data, stage = "a", y, outliers = NULL)
}
\arguments{
\item{data}{Data frame containing movie rating data}

\item{stage}{Character indicating the stage for which the data is prepared
a - used for the exploratory data analysis
m - used for modeling stage}

\item{y}{Character string containing the name of the response variable}

\item{outliers}{Vector of observations to remove}
}
\value{
List with two data frames, independent variables and the dependent
variable.
}
\description{
\code{process} Prepares the data for the bivariate analysis stage. Filters
out non-essential data, outliers and splits the data set into a set of
dependent variables and set with the response variable.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{challenger}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{createDependent}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
