% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionAnalysis.R
\name{regressionAnalysis}
\alias{regressionAnalysis}
\title{regressionAnalysis}
\usage{
regressionAnalysis(mod, mName, yVar, yLab)
}
\arguments{
\item{mod}{Linear model}

\item{mName}{Capitalized character string for the name of the linear model}

\item{yVar}{Character string containing the name of the dependent variable}

\item{yLab}{Capitalized character string for the dependent variable}
}
\value{
analysis plots, summary statistics, test results and interpretive text
}
\description{
\code{regressionAnalysis} Performs regression analysis
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{correlate}},
  \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
