% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(mdb)
}
\arguments{
\item{mdb}{Data frame containing movie information}
}
\description{
\code{univariate} Performs univariate analysis of variables
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{createDependent}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
