% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateQual.R
\name{univariateQual}
\alias{univariateQual}
\title{univariateQual}
\usage{
univariateQual(data, xLab)
}
\arguments{
\item{data}{Single column data frame containing the categorical variable}

\item{xLab}{Capitalized character string for the variable name or label}
}
\value{
analysis List containing:
 1. Contingency table
 2. Frequency Proportion Barplot
}
\description{
\code{univariateQual} Performs univariate analysis on a single qualitative
or categorical variable. The function returns a contingency table as well
as a stacked bar plot showing frequencies and percentages.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{challenger}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{createDependent}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{prediction}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{regressionAnalysis}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
