% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateQuant.R
\name{univariateQuant}
\alias{univariateQuant}
\title{univariateQuant}
\usage{
univariateQuant(data, yLab, units = NULL)
}
\arguments{
\item{data}{Single column data frame containing the quantitative variable}

\item{yLab}{Capitalized character string for the variable name or label}

\item{units}{Character string indicating the units associated with the quantitative variable.}
}
\value{
analysis List containing summary statistics, plots and intepretation text.
}
\description{
\code{univariateQuant} Performs univariate analysis on a single quantitative
variable.  The function returns descriptive statistics, a histogram
and a box plot to check outliers.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{evalForwardModel}},
  \code{\link{forward}}, \code{\link{prediction}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
