% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R, R/correlate.R
\name{correlate}
\alias{correlate}
\alias{correlate}
\title{correlate}
\usage{
correlate(x, y, yLab = NULL, plot = FALSE)

correlate(x, y, yLab = NULL, plot = FALSE)
}
\arguments{
\item{x}{Data frame containing the explanatory variable data}

\item{y}{Data frame containing the response data}

\item{yLab}{Character string containing label for the response variable}

\item{plot}{Logical indicating whether to include a correlation plot}

\item{x}{Data frame containing the explanatory variable data}

\item{y}{Data frame containing the response data}

\item{yLab}{Character string containing label for the response variable}
}
\description{
\code{correlate} Given two data frames, one containing data for the explanatory
variables (x data frame) and another containing the data for the response
(y data frame), this function conducts correlation tests and produces
correlation plots.

\code{correlate} Performs the correlate analysis between a set of
independent variables (x) and a dependent variable (y).  Conducts a
correlation test and renders correlation and scatter plots.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{evaluate}},
  \code{\link{forward}}, \code{\link{fullModel}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}

Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{evaluate}},
  \code{\link{forward}}, \code{\link{fullModel}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
