% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{prediction}
\usage{
prediction(mod, mName, case, unit, conf = 0.9)
}
\arguments{
\item{mod}{Linear regression model}

\item{mName}{Character string containing the model name}

\item{case}{Data frame containing single case to predict}

\item{unit}{Character string indicating the unit for the value being predicted (w/o log)}

\item{conf}{Numeric indicator (0 to 1) of the confidence level for the prediction interval.}
}
\value{

}
\description{
\code{prediction} Performs prediction on an randomly sampled individual observation
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{challenger}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{createDependent}},
  \code{\link{evalForwardModel}}, \code{\link{forward}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
