% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{prediction}
\usage{
prediction(mod, data)
}
\arguments{
\item{mod}{Linear regression model}

\item{data}{Data frame containing the movie data}
}
\value{

}
\description{
\code{prediction} Performs prediction on an randomly sampled individual observation
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{evalForwardModel}},
  \code{\link{forward}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
