% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(movies, mdb2)
}
\arguments{
\item{movies}{Data frame containing the movies data set}

\item{mdb2}{Data set of 100 randomly selected observations from the movies data set, in which total box office revenue was added.}
}
\description{
\code{preprocess} Performs preprocessing of data for analysis
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{correlate}},
  \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{evaluate}},
  \code{\link{forward}}, \code{\link{fullModel}},
  \code{\link{multivariateQuant}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
