% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{process}
\usage{
process(mdb)
}
\arguments{
\item{mdb}{Data frame containing movie rating data}
}
\value{
df Data frame of clean movie data
}
\description{
\code{process} Prepares the data for the bivariate analysis stage
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{bivariate}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}},
  \code{\link{regressionAnalysis}}, \code{\link{response}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
