% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{process}
\usage{
process(data, group = "a", y, outliers = NULL)
}
\arguments{
\item{data}{Data frame containing movie rating data}

\item{group}{Character indicating the data group to return.
a - Most conservative. Includes almost all variables
b <- Moderate data set with some data variables remomved
c >- Liberat data set focusing only on variables most predictive of log IMDB votes}

\item{y}{Character string containing the name of the response variable}

\item{outliers}{Vector of observations to remove}
}
\value{
List with two data frames, independent variables and the dependent
variable.
}
\description{
\code{process} Prepares the data for the bivariate analysis stage. Filters
out non-essential data, outliers and splits the data set into a set of
dependent variables and set with the response variable.
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{backwardElimination}}, \code{\link{back}},
  \code{\link{comparePredictions}},
  \code{\link{correlate}}, \code{\link{evalForwardModel}},
  \code{\link{forward}}, \code{\link{prediction}},
  \code{\link{preprocess}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
