% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(mdb)
}
\arguments{
\item{dataSets}{List containing the two movie data bases}
}
\description{
\code{univariate} Performs univariate analysis of variables
}
\seealso{
Other movies functions: \code{\link{associate}},
  \code{\link{back}}, \code{\link{correlate}},
  \code{\link{evalBackwardModel}},
  \code{\link{evalForwardModel}}, \code{\link{evaluate}},
  \code{\link{forward}}, \code{\link{fullModel}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
